Restructure folders and add a mobile version of the search bar
